Move deleter out of the helper package 🌿
